Table fixes and changes
Changed the onClick event in the rows of the table, from icon to button (before, it did register only if you clicked the icon, now it works on the whole button). Moved the table that contains the waiting to join miners from the MiningPool component to the VotingJoiners component (in it's correct place)
